app/requirements.txt
aiohttp
discord.py
nltk
numpy
python-dotenv
